feat: add RouterDecorator to Storybook preview
