+++ packages/ui/src/atoms/backlink.tsx
@@ -1,0 +1,19 @@
+import { Link } from '@tanstack/react-router';
+import { MdKeyboardArrowLeft } from 'react-icons/md';
+
+interface BackLinkProps {
+  to?: string;
+  label?: string;
+  className?: string;
+}
+
+export const BackLink = ({
+  to,
+  label,
+  className = 'flex items-center subtitle-large-med-20px md:display-large text-darkOrange-5 hover:text-white mb-[30px]',
+}: BackLinkProps) => (
+  <Link to={to} className={className}>
+    <MdKeyboardArrowLeft className="size-[18px] md:size-12" />
+    <span className="ml-1">{label}</span>
+  </Link>
+);
